requirements.txt
requests>=2.31.0
feedparser>=6.0.10
PyYAML>=6.0.1
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2
rapidfuzz>=3.9.0
python-dotenv>=1.0.1
